pyproject.toml
[tool.poetry]
name = "drfcommon"
version = "0.1.5"
description = "a common tools for drf"
authors = ["huoyinghui <hyhlinux@163.com>"]
homepage = "https://github.com/pydtools/drfcommon/"
repository = "https://github.com/pydtools/drfcommon/"
documentation = "https://github.com/pydtools/drfcommon/wiki"
include = ["CHANGELOG.md"]
[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2 || ^3.0"
djangorestframework = "^3.12.2"
drfhumplib = "^0.1.4.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/pydtools/pytimetools/issues"
"github" = "https://github.com/pydtools/pytimetools/"